+++ src/client.ts
@@ -93,6 +93,7 @@ import {
   GetImportResponse,
   GetMessageAPIResponse,
   GetRateLimitsResponse,
+  GetUnreadCountAPIResponse,
   ListChannelResponse,
   ListCommandsResponse,
   ListImportsPaginationOptions,
@@ -1584,17 +1585,6 @@ async connect() {
       channels.push(c);
     }
 
-    if (!offlineMode) {
-      // If the channels are coming from server, then clear out the
-      // previously help offline channels.
-      for (const key in this.activeChannels) {
-        const channel = this.activeChannels[key];
-        if (channel.offlineMode) {
-          delete this.activeChannels[key];
-        }
-      }
-    }
-
     return channels;
   }
 
@@ -1683,6 +1673,10 @@ async connect() {
       this.baseURL + '/devices',
       userID ? { user_id: userID } : {},
     );
+  }
+
+  async getUnreadCount(userID?: string) {
+    return await this.get<GetUnreadCountAPIResponse>(this.baseURL + '/unread', userID ? { user_id: userID } : {});
   }
 
   /**